Add files via upload
ATtiny85 is a small and low-power microcontroller designed by Atmel, now owned by Microchip Technology. It is part of the AVR family of microcontrollers and is commonly used in projects that require a small form factor, low power consumption, and basic functionality. It has 8 KB of flash memory for program storage, 512 bytes of EEPROM for non-volatile data storage, and 512 bytes of SRAM for temporary data storage. It also has 6 general-purpose I/O pins, including PWM capabilities, and can be programmed using a variety of development tools and programming languages, including the Arduino IDE. Due to its small size and low cost, it is a popular choice for DIY electronics projects, such as LED lighting, robotics, and sensor applications.